requirements copy.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
alembic=1.4.1=pypi_0
appnope=0.1.2=py38hecd8cb5_1001
argon2-cffi=20.1.0=py38h9ed2024_1
async_generator=1.10=pyhd3eb1b0_0
attrs=20.3.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=3.3.0=pyhd3eb1b0_0
botorch=0.4.0=pypi_0
bzip2=1.0.8=h1de35cc_0
ca-certificates=2021.4.13=hecd8cb5_1
certifi=2020.12.5=py38hecd8cb5_0
cffi=1.14.5=py38h2125817_0
chardet=4.0.0=pypi_0
click=7.1.2=pypi_0
cliff=3.7.0=pypi_0
cloudpickle=1.6.0=pypi_0
cmaes=0.8.2=pypi_0
cmd2=1.5.0=pypi_0
colorama=0.4.4=pypi_0
colorlog=4.8.0=pypi_0
cycler=0.10.0=py38_0
databricks-cli=0.14.3=pypi_0
decorator=5.0.6=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
docker=4.4.4=pypi_0
entrypoints=0.3=pypi_0
ffmpeg=4.3=h0a44026_0
flask=1.1.2=pypi_0
freetype=2.10.4=ha233b18_0
gettext=0.19.8.1=hb0f4f8b_2
gitdb=4.0.7=pypi_0
gitpython=3.1.14=pypi_0
gmp=6.2.1=h23ab428_2
gnutls=3.6.5=h91ad68e_1002
gpytorch=1.4.0=pypi_0
greenlet=1.0.0=pypi_0
gunicorn=20.1.0=pypi_0
helper=2.5.0=pypi_0
idna=2.10=pypi_0
importlib-metadata=3.10.0=py38hecd8cb5_0
importlib_metadata=3.10.0=hd3eb1b0_0
intel-openmp=2019.4=233
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.22.0=py38h01d92e1_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.3=pypi_0
itsdangerous=1.1.0=pypi_0
jedi=0.17.0=py38_0
jinja2=2.11.3=pyhd3eb1b0_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=he5867d9_2
jsonschema=3.2.0=py_2
jupyter=1.0.0=pypi_0
jupyter-console=6.4.0=pypi_0
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.7.1=py38hecd8cb5_0
jupyterlab-widgets=1.0.0=pypi_0
jupyterlab_pygments=0.1.2=py_0
kiwisolver=1.3.1=py38h23ab428_0
lame=3.100=h1de35cc_0
lcms2=2.11=h92f6f08_0
libcxx=10.0.0=1
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libiconv=1.16=h1de35cc_0
libpng=1.6.37=ha441bb4_0
libsodium=1.0.18=h1de35cc_0
libtiff=4.2.0=h87d7836_0
libuv=1.40.0=haf1e3a3_0
libwebp-base=1.2.0=h9ed2024_0
llvm-openmp=10.0.0=h28b9765_0
lz4-c=1.9.3=h23ab428_0
mako=1.1.4=pypi_0
markupsafe=1.1.1=py38h1de35cc_1
matplotlib=3.3.4=py38hecd8cb5_0
matplotlib-base=3.3.4=py38h8b3ea08_0
mistune=0.8.4=py38h1de35cc_1001
mkl=2019.4=233
mkl-service=2.3.0=py38h9ed2024_0
mkl_fft=1.3.0=py38ha059aab_0
mkl_random=1.1.1=py38h959d312_0
mlflow=1.15.0=pypi_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.0.7=py38_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.2=h0a44026_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
nettle=3.4.1=h3018a27_0
ninja=1.10.2=py38hf7b0b51_0
notebook=6.3.0=py38hecd8cb5_0
numpy=1.19.2=py38h456fd55_0
numpy-base=1.19.2=py38hcfb5961_0
olefile=0.46=py_0
openh264=2.1.0=hd9629dc_0
openssl=1.1.1k=h9ed2024_0
optuna=2.6.0=pypi_0
packaging=20.9=pyhd3eb1b0_0
pandas=1.2.3=py38hb2f4e1b_0
pandoc=2.12=hecd8cb5_0
pandocfilters=1.4.3=py38hecd8cb5_1
parso=0.8.2=pyhd3eb1b0_0
pbr=5.5.1=pypi_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.1.2=py38h5270095_0
pip=21.0.1=py38hecd8cb5_0
plotly=4.14.3=pypi_0
prettytable=2.1.0=pypi_0
prometheus-flask-exporter=0.18.1=pypi_0
prometheus_client=0.10.0=pyhd3eb1b0_0
prompt-toolkit=3.0.17=pyh06a4308_0
protobuf=3.15.7=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pycparser=2.20=py_2
pygments=2.8.1=pyhd3eb1b0_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyperclip=1.8.2=pypi_0
pyrsistent=0.17.3=py38haf1e3a3_0
python=3.8.8=h88f2d9e_4
python-dateutil=2.8.1=pyhd3eb1b0_0
python-editor=1.0.4=pypi_0
pytorch=1.8.1=py3.8_0
pytorchtools=0.0.2=pypi_0
pytz=2021.1=pyhd3eb1b0_0
pyyaml=5.4.1=pypi_0
pyzmq=20.0.0=py38h23ab428_1
qtconsole=5.0.3=pypi_0
qtpy=1.9.0=pypi_0
querystring-parser=1.2.4=pypi_0
readline=8.1=h9ed2024_0
requests=2.25.1=pypi_0
retrying=1.3.3=pypi_0
scikit-learn=0.24.1=py38hb2f4e1b_0
scipy=1.6.2=py38h2515648_0
send2trash=1.5.0=pyhd3eb1b0_1
setuptools=52.0.0=py38hecd8cb5_0
six=1.15.0=py38hecd8cb5_0
sklearn=0.0=pypi_0
skorch=0.10.0=pypi_0
smmap=4.0.0=pypi_0
sqlalchemy=1.4.3=pypi_0
sqlite=3.35.2=hce871da_0
sqlparse=0.4.1=pypi_0
stevedore=3.3.0=pypi_0
tabulate=0.8.9=pypi_0
terminado=0.9.4=py38hecd8cb5_0
testpath=0.4.4=pyhd3eb1b0_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hb0a8c7a_0
torchvision=0.9.1=py38_cpu
tornado=6.1=py38h9ed2024_0
tqdm=4.59.0=pypi_0
traitlets=5.0.5=pyhd3eb1b0_0
typing_extensions=3.7.4.3=pyha847dfd_0
urllib3=1.26.4=pypi_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py38_1
websocket-client=0.58.0=pypi_0
werkzeug=1.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=pypi_0
xz=5.2.5=h1de35cc_0
zeromq=4.3.4=h23ab428_0
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.5=h41d2c2f_0
